restructuring the reportserver to have same structure as the rest of the project.
