+++ reportserver/manager/dateTimeUtility.py
@@ -33,7 +33,7 @@ Last Revised: 28 February, 2016
 
 import datetime
 
-from manager.UnitOfMeasure import UnitOfMeasure
+from reportserver.manager import UnitOfMeasure
 
 
 # Return the date for how far back to query DB.
