+++ reportserver/server/main.py
@@ -1,0 +1,4 @@
+from reportserver.server.SimpleHttpServer import SimpleHttpServer
+
+server = SimpleHttpServer()
+server.setupAndStart();
