--- reportserver/src/main/README.md
@@ -1,1 +1,0 @@
-holds the source code for the honeypot
