--- reportserver/src/main/server/main.py
@@ -1,4 +1,0 @@
-from server.SimpleHttpServer import SimpleHttpServer
-
-server = SimpleHttpServer()
-server.setupAndStart();
